Made some small reorganization of the file structure
