+++ client/src/components/Pages/Disclaimer/index.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import debounce from 'lodash.debounce';
 import { Link } from 'react-router-dom';
 
-import Animation from '../Common/DotsAnimation';
-import { HOME } from '../../constants/routes';
+import Animation from '../../Common/DotsAnimation';
+import { HOME } from '../../../constants/routes';
 
 import './styles.css';
