+++ client/src/components/Pages/Education/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import EmptyView from '../Common/EmptyView';
+import EmptyView from '../../Common/EmptyView';
 
 const Education = () => (
   <div>
