+++ client/src/components/Pages/Home/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import EmptyView from '../Common/EmptyView';
+import EmptyView from '../../Common/EmptyView';
 
 const DESCRIPTION = `
   This is home
